% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpc.names}
\alias{cpc.names}
\title{CPC Names}
\format{An object of class \code{data.frame} with 400 rows and 3 columns.}
\source{
Global Trade Alert
}
\usage{
cpc.names
}
\description{
Dataframe containing CPC Names, sector codes and levels
}
\keyword{datasets}
