% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpc.to.hs}
\alias{cpc.to.hs}
\title{CPC to HS}
\format{An object of class \code{data.frame} with 5205 rows and 2 columns.}
\source{
Global Trade Alert
}
\usage{
cpc.to.hs
}
\description{
Dataframe containing CPC and HS12 code correspondance
}
\keyword{datasets}
