\name{gta_cpc_code_check}
\alias{gta_cpc_code_check}
\title{Check whether your CPC codes are consistent with CPC 2.1.}
\description{This function checks whether a vector of CPC codes is consistent with CPC 2.1 and returns it as a vector of 3-digit level codes.}
\usage{}
\arguments{
\item{codes}{Supply the CPC codes you want to check. Values with 2 or more digits are allowed.}
}
\details{}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}
