\name{gta_cpc_to_hs}
\alias{gta_cpc_to_hs}
\title{Convert CPC 2.1 codes into HS 2012 (6-digit level).}
\description{This function converts vectors of CPC 2.1 codes (any granularity) into a vector of 6-digit HS 2012 codes.}
\usage{}
\arguments{
\item{codes}{Supply the CPC codes you want to convert. Values with 2 or more digits are allowed.}
}
\details{}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}
