\name{gta_hs_code_check}
\alias{gta_hs_code_check}
\title{Check whether your HS codes are consistent with HS 2012.}
\description{This function checks whether a vector of HS codes is consistent with HS 2012 and returns it as a vector of 6-digit level HS codes.}
\usage{}
\arguments{
\item{codes}{Supply the HS codes you want to check. Values with 2 or more digits are allowed.}
}
\details{}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}
