\name{gta_hs_to_cpc}
\alias{gta_hs_to_cpc}
\title{Convert HS 2012 codes into CPC 2.1 codes (3-digit level).}
\description{This function converts vectors of HS 2012 codes (any granularity) into a vector of 3-digit CPC 2.1 codes.}
\usage{}
\arguments{
\item{codes}{Supply the HS codes you want to convert. Values with 2, 4, 6 or 8 digits are allowed.}
}
\details{}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}
