\name{gta_cpc_to_hs}
\alias{gta_cpc_to_hs}
\title{Convert CPC 2.1 codes into HS 2012 (6-digit level).}
\description{This function converts vectors of CPC 2.1 codes (any granularity) into a vector of 6-digit HS 2012 codes.}
\usage{}
\arguments{
\item{codes}{Supply the CPC codes you want to convert. Values with 2 or more digits are allowed.}
}
\details{}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}

\name{gta_cpc_code_check}
\alias{gta_cpc_code_check}
\title{Check whether your CPC codes are consistent with CPC 2.1.}
\description{This function checks whether a vector of CPC codes is consistent with CPC 2.1 and returns it as a vector of 3-digit level codes.}
\usage{}
\arguments{
\item{codes}{Supply the CPC codes you want to check. Values with 2 or more digits are allowed.}
}
\details{}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}


\name{gta_hs_to_cpc}
\alias{gta_hs_to_cpc}
\title{Convert HS 2012 codes into CPC 2.1 codes (3-digit level).}
\description{This function converts vectors of HS 2012 codes (any granularity) into a vector of 3-digit CPC 2.1 codes.}
\usage{}
\arguments{
\item{codes}{Supply the HS codes you want to convert. Values with 2, 4, 6 or 8 digits are allowed.}
}
\details{}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}


\name{gta_hs_code_check}
\alias{gta_hs_code_check}
\title{Check whether your HS codes are consistent with HS 2012.}
\description{This function checks whether a vector of HS codes is consistent with HS 2012 and returns it as a vector of 6-digit level HS codes.}
\usage{}
\arguments{
\item{codes}{Supply the HS codes you want to check. Values with 2 or more digits are allowed.}
}
\details{}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}
