\name{gta_library_update}
\alias{gta_library_update}
\title{Get the latest version from GitHub.}
\description{Syncs your local library with our latest GTA GitHub release.}
\usage{}
\arguments{
}
\details{}
\value{Be up to date with our latest functions.}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}
