\name{gta_un_code_vector}
\alias{gta_un_code_vector}
\title{Create a vector of UN country codes}
\description{Returns a vector of codes based on country or group names.}
\usage{}
\arguments{
\item{countries}{Vector with either country names, country group names or UN codes. Do not mix UN codes with country or group names. Please use function also for UN codes to ensure the ones you supply are consistent with the ones used by the GTA}
\item{role}{This parameter tailors the error message e.g. "Unkown ROLE country value ...". Only useful if command is run for more than one set of countries.}
}
\details{}
\value{A vector of UN country codes that is consistent with those used by the GTA.}
\references{www.globaltradealert.org}
\author{Global Trade Alert}
\examples{}
