% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean country names.R
\name{gta_clean_country_names}
\alias{gta_clean_country_names}
\title{Convert country names to GTA nomenclature}
\usage{
gta_clean_country_names(
    country,
    conversionTable = FALSE
)
}
\arguments{
\item{country}{a vector of country names}

\item{conversionTable}{set to TRUE if the function should return a conversion table
which shows how each unique value in `country` is converted by the function}
}
\description{
\code{gta_clean_country_names()} takes a vector of country names
as an input and converts them to the country names used in the GTA database.
}
\details{
If a country cannot or cannot uniquely be matched to one country in the gta nomenclature,
NA is returned.
If you want a table with unique country names from the input vector, the name
they were converted to and information on whether the function changed the name, set conversionTable
to true. This will output a tibble with these three columns.
}
\references{
The function uses gtalibrary::country_regex which contains a regex for each
country which inputs are matched against.
}
