% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cloud update.R
\name{gta_cloud_update}
\alias{gta_cloud_update}
\title{Update your local copy of our data set.}
\usage{
gta_cloud_update()
}
\value{
Be up to date with our latest data.
}
\description{
Update your local copy of our data set.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
