% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta collect dom.R
\name{gta_collect_dom}
\alias{gta_collect_dom}
\title{Extract the complete DOM of an XML file.}
\usage{
gta_collect_dom(node = NULL, dom.pos = c())
}
\arguments{
\item{node}{Provide an xmlRoot object.}

\item{dom.position}{Specify the starting level for your DOM. Default is 1 (=root).}
}
\description{
Extract the complete DOM of an XML file.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
