% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta collect url.R
\name{gta_collect_url}
\alias{gta_collect_url}
\title{Collects a URL and stores it locally using file name. Websites are printed into PDF, URLS to files are downloaded.}
\usage{
gta_collect_url(
  url = NULL,
  file.name = NULL,
  store.path = NULL,
  add.timestamp = TRUE,
  update.file.name = TRUE,
  js.path = "setup/rasterize.js",
  phantom.path = "~/bin/phantomjs"
)
}
\value{
The complete file name of the collected url (incl. file format).
}
\description{
Collects a URL and stores it locally using file name. Websites are printed into PDF, URLS to files are downloaded.
}
\references{
www.globaltradealert.org
}
\author{
GTA
}
