% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta colour palette.R
\name{gta_colour_palette}
\alias{gta_colour_palette}
\title{GTA Colour Palette .}
\usage{
gta_colour_palette()
}
\description{
Access any colour by calling e.g. gta_colour$green[1]. Every colour comes in 4 predefined shades (index 1 to 4 or dark to light).
}
\details{
'qualitative' contains 8 neutral colours. Shade functions can be filled with any numeric value indicating the number of shades needed.

Applicable values are: green, liberalising, green.light, red, harmful, red.light, amber, blue, brown, turquoise, desert, grey, blue.complement, qualitative, green.shades(), red.shades(), amber.shades(), blue.shades(), turquoise.shades(), desert.shades(), brown.shades(), grey.shades().
}
