% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta country code vector.R
\name{gta_country_code_vector}
\alias{gta_country_code_vector}
\title{Create a vector of UN country codes}
\usage{
gta_country_code_vector(countries, destination.code = "UN")
}
\arguments{
\item{countries}{Vector with either country names, country group names or UN codes. Do not mix UN codes with country or group names. Please use function also for UN codes to ensure the ones you supply are consistent with the ones used by the GTA.}

\item{destination.code}{Specify whether which destination code you want to receive. Options are UN, GTA or ISO3. Default is UN.}
}
\value{
A vector of UN country codes that is consistent with those used by the GTA.
}
\description{
Returns a vector of codes based on country or group names.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
