% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cpc code check.R
\name{gta_cpc_code_check}
\alias{gta_cpc_code_check}
\title{Check whether your CPC codes are consistent with CPC 2.1.}
\usage{
gta_cpc_code_check(codes)
}
\arguments{
\item{codes}{Supply the CPC codes you want to check. Only 3-digit level CPC codes are allowed. If mixing 2-digit and 3-digit CPC sector codes, place 2-digit codes inside the gta_cpc_code_expand function.}
}
\description{
This function checks whether a vector of CPC codes is consistent with CPC 2.1 and returns it as a vector of 3-digit level codes.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
