% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cpc code expand.R
\name{gta_cpc_code_expand}
\alias{gta_cpc_code_expand}
\title{Expands 2nd level CPC codes to the 3rd level.}
\usage{
gta_cpc_code_expand(codes)
}
\arguments{
\item{codes}{Supply the 2nd level CPC codes you want to expand.}
}
\description{
This function takes in CPC 2nd level codes and outputs all corresponding 3rd level codes in a vector.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
