% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cpc names.R
\name{gta_cpc_names}
\alias{gta_cpc_names}
\title{Create a dataframe with CPC names out of a list of CPC codes.}
\usage{
gta_cpc_names(codes, level = NULL)
}
\arguments{
\item{codes}{Supply the CPC codes of which you want to get the name.}

\item{level}{Specify the level of the codes provided (2 or 3).}
}
\description{
This function returns the names of CPC codes.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
