% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cumulative counts.R
\name{gta_cumulative_counts}
\alias{gta_cumulative_counts}
\title{GTA in force interventions counter}
\usage{
gta_cumulative_counts(
  counts.by = "quarter",
  coverage.period = NULL,
  data.path = "data/master_plus.Rdata",
  gta.evaluation = NULL,
  affected.flows = NULL,
  implementing.country = NULL,
  keep.implementer = TRUE,
  affected.country = NULL,
  keep.affected = NULL,
  incl.affected.strictness = "ONEPLUS",
  keep.others = TRUE,
  nr.affected = c(0, 999),
  nr.affected.incl = "ALL",
  announcement.period = NULL,
  implementation.period = NULL,
  keep.implementation.na = NULL,
  revocation.period = NULL,
  keep.revocation.na = TRUE,
  submission.period = NULL,
  in.force.on.date = Sys.Date(),
  keep.in.force.on.date = "any",
  intervention.types = NULL,
  keep.type = NULL,
  mast.chapters = NULL,
  keep.mast = NULL,
  implementation.level = NULL,
  keep.level = NULL,
  eligible.firms = NULL,
  keep.firms = NULL,
  cpc.sectors = NULL,
  keep.cpc = NULL,
  hs.codes = NULL,
  keep.hs = NULL,
  intervention.ids = NULL,
  keep.interventions = NULL,
  lag.adjustment = NULL,
  reporting.period = NULL,
  df.name = "master.sliced",
  pc.name = "parameter.choice.slicer",
  xlsx = FALSE,
  output.path = NULL
)
}
\arguments{
\item{counts.by}{Specify whether to count by month, quarter or year.'.}

\item{coverage.period}{Specify the range of years of interest.'.}
}
\value{
Outputs in force interventions by different given periods.
}
\description{
Computes number of interventions in force.
}
\details{
all gta_data_slicer parameters are permissible.'.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
