% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta data download.R
\name{gta_data_download}
\alias{gta_data_download}
\title{Download the latest copy of the GTA database.}
\usage{
gta_data_download(download.folder = NULL, load.data = FALSE)
}
\arguments{
\item{download.folder}{Set the download folder. Default is current working directory.}

\item{load.data}{Loads the downloaded data into the current environment. Default is FALSE.}
}
\description{
This function downloads the latest copy of the full GTA database export and stores it locally.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
