% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta get imf data.R
\name{gta_get_imf_data}
\alias{gta_get_imf_data}
\title{A wrapper for the IMFData function}
\usage{
gta_get_imf_data(
  start.date = NULL,
  end.date = NULL,
  frequency = NULL,
  series = NULL,
  countries = NULL
)
}
\arguments{
\item{start.date}{Provide the period start date in R format ('YYYY-mm-dd').}

\item{end.date}{Provide the period end date in R format ('YYYY-mm-dd').}

\item{series}{What time series do you want? Options: 'fx' for USD exchange rates.}

\item{countries}{What countries do you want? Permissiable options are 'all' plus GTA names and ISO3 codes.}

\item{fx.frequency}{Provide time series frequency e.g. ('D','M','Q','A')}
}
\description{
A wrapper for the IMFData function
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
