% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta hs code check.R
\name{gta_hs_code_check}
\alias{gta_hs_code_check}
\title{Check whether your HS codes are consistent with HS 2012.}
\usage{
gta_hs_code_check(codes)
}
\arguments{
\item{codes}{Supply the HS codes you want to check. Values with 2 or more digits are allowed. Values with more than 6 digits will be limited to 6. Please input the codes as integer.}
}
\description{
This function checks whether a vector of HS codes is consistent with HS 2012 and returns it as a vector of 6-digit level HS codes.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
