% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta hs code finder.R
\name{gta_hs_code_finder}
\alias{gta_hs_code_finder}
\title{Find HS codes for product names based on the search engines from EU customs and Eurostat.}
\usage{
gta_hs_code_finder(products, sources = c("hs.descriptions", "eurostat",
  "eu.customs", "zauba", "e.to.china", "google", "hsbianma", "eximguru",
  "cybex"))
}
\arguments{
\item{products}{A vector of product names (in characters).}

\item{source}{Define which source(s) to check for HS codes (hs.descriptions, eurostat, eu.customs, zauba, e.to.china, google,  hsbianma, eximguru). Default is 'all'.}
}
\value{
A data frame including the product names, HS codes and sources plus a vector of search terms that resulted in error messages.
}
\description{
This function looks for HS codes using various search engines.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
