% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta hs names.R
\name{gta_hs_names}
\alias{gta_hs_names}
\title{Create a dataframe with HS names out of a list of HS codes.}
\usage{
gta_hs_names(codes)
}
\arguments{
\item{codes}{Supply the HS codes of which you want to get the name.}
}
\description{
This function returns the names of HS codes on a 6 digit level.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
