% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta hs vintage converter.R
\name{gta_hs_vintage_converter}
\alias{gta_hs_vintage_converter}
\title{Convert HS codes of unkown vintage into HS 2012.}
\usage{
gta_hs_vintage_converter(codes, origin = "best")
}
\arguments{
\item{codes}{Supply the HS codes you want to check. Values with 2 or more digits are allowed. Values with more than 6 digits will be limited to 6. Please input the codes as integer.}

\item{origin}{State the origin vintage, if you know it (options: '2002','2007','2012','2017','best', 'any'). Where 'best' selects the HS vintage with the most matches and 'any' selects HS codes from multiple origin vintages. Default is 'best'.}
}
\description{
Returns it as a vector of 6-digit level HS 2012 codes.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
