% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta ics to hs.r
\name{gta_ics_to_hs}
\alias{gta_ics_to_hs}
\title{Convert ICS codes to 4 digit HS codes}
\usage{
gta_ics_to_hs(codes, as_list = FALSE, message = TRUE)
}
\arguments{
\item{codes}{Supply the ICS codes you want to check. Only codes in 5 digit format are converted (eg. 67120 / 67.120). Shorter codes are ignored, longer codes are truncated to 5 digits}

\item{as_list}{returns the results as a list with the same length as the codes input vector, FALSE by default}

\item{message}{Prints conversion results if TRUE}
}
\description{
Convert ICS codes to 4 digit HS codes
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
