% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta intervention duration.R
\name{gta_intervention_duration}
\alias{gta_intervention_duration}
\title{GTA intra-year duration calculator}
\usage{
gta_intervention_duration(data.path = "data/master_plus.Rdata",
  is.data.frame = FALSE, years = NULL, current.year.todate = TRUE,
  df.name = "intervention.duration",
  pc.name = "parameter.choice.duration")
}
\arguments{
\item{data.path}{Specifies where the GTA data file is located (Default: 'data/master_plus.Rdata'). Set to 'online' to download the latest copy.}

\item{is.data.frame}{Specify if the data path is a data frame in the current environment. Default is FALSE. If you supply a data frame, the first three columns need to be (1) intervention ID, (2) implementation date, (3) removal date. In that order.}

\item{years}{The calendar years for which to calculate the shares. Calculation includes interventions based on enforcement status, not implementation date i.e. if you start in 2010, this function will also work with interventions implemneted in 2009 but still in force in 2010. Specify as c(start.year, end.year). Default is c(2008,CURRENT.YEAR).}

\item{current.year.todate}{Should the duration for the current year, if included, be calculated as 'share of year to date' (TRUE) or 'share of entire current year' (FALSE). Default is TRUE.}

\item{df.name}{Set the name of the generated result data frame. Default is intervention.duration.}

\item{pc.name}{Set the name of the generated parameter choice data frame. Default is parameter.choice.duration.}
}
\value{
Output is two data frames. First data frame includes the share of each year that in intervention was in force. Second data frame states parameter choices.
}
\description{
Computes the share of each calendar year that each intervention was in force.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
