% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta logical check.r
\name{gta_logical_check}
\alias{gta_logical_check}
\title{Check whether function arguments fulfill a logical conditoion}
\usage{
gta_logical_check(
     check_arg,
     check_function,
     error_msg = NULL
)
}
\arguments{
\item{check_arg}{The argument / variable that you want to check}

\item{check_function}{A function that takes a scalar `check_arg` as an input and returns a boolean
If you pass internal functions (eg. \code{is.numeric}), pass them without brackets.
You can also pass anonymous functions (eg. \code{\(x)}  \code{is.numeric(x) & length(x) >= 10)})}

\item{error_msg}{Message that is printed if check_function returns FALSE. Glue Syntax (ie. passing of an expression
which should be evaluated with {} is supported)}
}
\description{
Check whether function arguments fulfill a logical conditoion
}
\examples{
# 1.) you want to make sure that a is numeric, and b is numeric and larger than 0
example_function(a, b){
gta_is_logical(a, is.numeric, error_msg = "{.var a} must be logical")
gta_is_logical(b, \(x) is.numeric(x) & x > 0, error_msg = "{.var b} must be larger than 10")
return(log(a) + b)
}
}
