% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta plot 3d map.R
\name{gta_plot_3d_map}
\alias{gta_plot_3d_map}
\title{Create 3D world map.}
\usage{
gta_plot_3d_map(
  data = NULL,
  countries = NULL,
  value.x = NULL,
  value.y = NULL,
  value.x.breaks = 4,
  value.y.breaks = 4,
  x.high.y.low = gta_colour$blue[1],
  x.high.y.high = gta_colour$blue[4],
  x.low.y.high = gta_colour$red[4],
  x.low.y.low = gta_colour$red[1],
  x.axis.title = NULL,
  y.axis.title = NULL,
  x.axis.labels = NULL,
  y.axis.labels = NULL,
  legend.title = NULL,
  save.eps = FALSE,
  save.png = FALSE,
  save.path = "/",
  save.name = "3D-Map"
)
}
\arguments{
\item{data}{Specify which dataset contains the data. The dataset must have a column with UN codes, and a column for x and y values each.}

\item{countries}{Specify the name of the country UN codes column as string.}

\item{value.x}{Specify the name of the x value column as string.}

\item{value.y}{Specify the name of the y value column as string.}

\item{value.x.breaks}{Specify in how many tiles the x data should be splitted und numeric form. Default is 4.}

\item{value.y.breaks}{Specify in how many tiles the y data should be splitted und numeric form. Default is 4.}

\item{x.high.y.low}{Specify the color of the bottom right matrix corner. Applicable is any colour in the gta colour palette or a HEX colour code. Default: blue[1].}

\item{x.high.y.high}{Specify the color of the top right matrix corner. Applicable is any colour in the gta colour palette or a HEX colour code. Default: blue[4].}

\item{x.low.y.high}{Specify the color of the top left matrix corner. Applicable is any colour in the gta colour palette or a HEX colour code. Default: red[4].}

\item{x.low.y.low}{Specify the color of the bottom left matrix corner. Applicable is any colour in the gta colour palette or a HEX colour code. Default: red[1].}

\item{x.axis.title}{Specify the name of the colour matrix x-axis. Default: NULL.}

\item{y.axis.title}{Specify the name of the colour matrix y-axis. Default: NULL.}

\item{x.axis.labels}{Specify the labels along the colour matrix x-axis. If labels are provided, please specify as many labels as there are breaks, empty labels applicable. Default: NULL.}

\item{y.axis.labels}{Specify the labels along the colour matrix y-axis. If labels are provided, please specify as many labels as there are breaks, empty labels applicable. Default: NULL.}

\item{legend.title}{Specify the title of the colour matrix. It is advisable to use the title for axis explanations, as it takes up less space than two additional axis titles. Default: NULL.}

\item{save.eps}{Specify whether you want to save eps file or not ('TRUE' and 'FALSE'). Plot will be returned as an object and can also be saved manually. Default: FALSE.}

\item{save.png}{Specify whether you want to save png file or not ('TRUE' and 'FALSE'). Plot will be returned as an object and can also be saved manually. Default: FALSE.}

\item{save.path}{Specify the path the plots will be saved to. Default: Working directory.}

\item{save.name}{Specify the name of the files to be saved.}
}
\description{
This function creates a world map with 3-dimensional values.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
