% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta plot map.R
\name{gta_plot_map}
\alias{gta_plot_map}
\title{Create ggplot maps.}
\usage{
gta_plot_map(data = NULL, countries = NULL, value = NULL,
  value.center = NULL, colour.low = blue[4], colour.high = blue[1],
  colour.breaks = waiver(), colour.legend.title = NULL)
}
\arguments{
\item{data}{Specify the dataframe with the relevant columns.}

\item{countries}{Specify the name of the column containing the countries.}

\item{value}{Specify the name of the column containing the value for each country.}

\item{value.center}{Specify the name of the column containing values for an area variable in the center of each country.}

\item{colour.low}{Specifiy a color from the gta colour palette for the low part of the gradient. Default: GTA dark blue}

\item{colour.high}{Specify a color from the gta colour palette for the high part of the gradient. Default: GTA light blue}

\item{colour.breaks}{Specify the breaks for the colour gradient. Default: ggplot automatic.}

\item{colour.legend.title}{Specify a title for the colour legend. Default: No title.}
}
\description{
This function facilitates map creation with ggplot.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
