% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta plot map df.R
\name{gta_plot_map_df}
\alias{gta_plot_map_df}
\title{Map geo data to a dataframe with un codes.}
\usage{
gta_plot_map_df(data = vdma.master, countries = NULL, values = NULL)
}
\arguments{
\item{data}{Specify the dataframe with the relevant columns.}

\item{countries}{Specify the name of the column containing the countries.}

\item{values}{Specify the name of the column containing the value for each country.}
}
\description{
This function facilitates map creation with ggplot.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
