% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta plot saver.R
\name{gta_plot_saver}
\alias{gta_plot_saver}
\title{Save png and eps plots together more easily.}
\usage{
gta_plot_saver(
  plot = NULL,
  path = "/",
  name = "plot",
  width = 21,
  height = NULL,
  png = T,
  eps = F,
  pdf = F,
  jpg = F,
  cairo_ps = F,
  aspect.ratio = 21/29.7
)
}
\arguments{
\item{plot}{Specify the plot to save.}

\item{path}{Specify the folder to save the plot into as string. Default: working directory.}

\item{name}{Input the name of the plot as string. Default: "plot".}

\item{width}{Takes in a specific plot height. If unspecified, height will be calculated through aspect.ratio.}

\item{png}{Set png output to FALSE. Default: TRUE.}

\item{eps}{Set cairo_ps output to TRUE if you want to generate an eps file with transparency levels. Default: FALSE.}

\item{aspect.ratio}{Define the aspect ratio of the plot as integer. Width = 1, height = 1*aspect.ratio. If height is defined, aspect.ratio does have no effect.}
}
\description{
This function saves ggplot objects as png and eps. Optimized for width of A4 paper, but can be changed accordingly.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
