% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta plot tile.R
\name{gta_plot_tile}
\alias{gta_plot_tile}
\title{Create ggplot maps.}
\usage{
gta_plot_tile(
  data = NULL,
  value.x = NULL,
  value.y = NULL,
  values = NULL,
  colour.low = gta_colour$blue[4],
  colour.high = gta_colour$blue[1],
  legend.title = NULL,
  title = NULL,
  x.axis.name = NULL,
  y.axis.name = NULL
)
}
\arguments{
\item{data}{Specify the dataframe with the relevant columns.}

\item{value.x}{Specify the name of the column containing the x-values.}

\item{value.y}{Specify the name of the column containing the y-values.}

\item{values}{Specify the name of the column containing the value for each country.}

\item{colour.low}{Specifiy a color from the gta colour palette for the low part of the gradient. Default: GTA dark blue}

\item{colour.high}{Specify a color from the gta colour palette for the high part of the gradient. Default: GTA light blue}

\item{legend.title}{Specify a title for the colour legend. Default: No title.}

\item{title}{Specify a title for the map plot. Default: No title.}

\item{x.axis.name}{Specify the x-axis label. Default: No label.}

\item{y.axis.name}{Specify the x-axis label. Default: No label.}
}
\description{
This function facilitates map creation with ggplot.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
