% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta pwd.R
\name{gta_pwd}
\alias{gta_pwd}
\title{GTA pwd function}
\usage{
gta_pwd(type = NULL)
}
\arguments{
\item{type}{If parameter type is specified, a single list will be returned. Otherwise all credentials will be returned. Default: All.}
}
\value{
Outputs either a list of credentials for a specified app or all credentials available for GTA.
}
\description{
Returns a list of GTA credentials.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
