% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta rbind.R
\name{gta_rbind}
\alias{gta_rbind}
\title{Rbind two dataframes with different columns.}
\usage{
gta_rbind(list)
}
\arguments{
\item{list}{Supply a list of dataframes. E.g. list = list(df1, df2, df3).}
}
\description{
This function rbinds two dataframes with a different set of columns and fills unmatched cells with NA.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
