% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta sop dataframes.R
\name{gta_sop_dataframes}
\alias{gta_sop_dataframes}
\title{Create a data frame that contains the names of the data frames used in the data dumps with their respective variables.}
\usage{
gta_sop_dataframes(
  aff.jur.df = F,
  dist.market.df = F,
  int.level.df = F,
  hs.codes.df = F,
  cpc.codes.df = F,
  firms.df = F,
  products.df = F,
  investigations.df = F,
  frameworks.df = F,
  related.acts.df = F,
  locations.df = F
)
}
\arguments{
\item{dataframes}{Specify any optional data frame you wish to use. Default FALSE.}
}
\value{
Dataframe with the necessary information for a given data dump.
}
\description{
Returns a data frame with the necessary information for the data dumps team.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
