% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta ted parse.R
\name{gta_ted_parse}
\alias{gta_ted_parse}
\title{Parsing TED XML files for GTA relevance}
\usage{
gta_ted_parse(dom.df = NULL, fx.df = NULL)
}
\arguments{
\item{dom.df}{Data frame including the DOM}

\item{fx.df}{Data frame including the FX rates}
}
\description{
Parsing TED XML files for GTA relevance
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
