% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta trade coverage.R
\name{gta_trade_coverage}
\alias{gta_trade_coverage}
\title{GTA trade coverage function}
\usage{
gta_trade_coverage(data.path = "data/master_plus.Rdata",
  data.path.coverage = "data/trade coverage/trade coverage base file.Rdata",
  gta.evaluation = NULL, affected.flow = NULL, importers = NULL,
  exporters = NULL, implementers = NULL, implementer.role = NULL,
  announcement.period = NULL, implementation.period = NULL,
  revocation.period = NULL, in.force.today = NULL,
  intervention.type = NULL, keep.type = NULL, mast.chapter = NULL,
  keep.mast = NULL, implementation.level = NULL, keep.level = NULL,
  eligible.firms = NULL, keep.firms = NULL, cpc.sectors = NULL,
  keep.cpc = NULL, hs.codes = NULL, keep.hs = NULL,
  intervention.ids = NULL, keep.interventions = NULL,
  lag.adjustment = NULL, rdata = FALSE, output.path = NULL,
  testmode = FALSE)
}
\arguments{
\item{data.path}{Specifies where the GTA data file is located (Default: 'data/master_plus.Rdata'). Set to 'online' to download the latest copy.}

\item{data.path.coverage}{Specifies where the GTA data file is located (Default: 'data/database replica'). Set to 'online' to download the latest copy.}

\item{gta.evaluation}{Specify what GTA evaluations to include. Default is 'any'. Permissible values are 'red', 'amber', 'green' or combinations thereof.}

\item{affected.flow}{Specify the direction of the trade flow that is affected. The point of view is from the implementing country. Default is 'any'. Permissible values are 'inward', 'outward', 'outward subsidy' or combinations thereof.}

\item{importers}{Takes in a list of country names, UN codes or country groups (g7, g20, eu28, ldc, au) to filter for importers in the sample. World will be added independent of the choice made. Default: All combinations.}

\item{exporters}{Takes in a list of country names, UN codes or country groups (g7, g20, eu28, ldc, au) to filter for exporters in the sample. World will be added independent of the choice made. Default: All combinations.}

\item{implementers}{Takes in a list of country names, UN codes or country groups (g7, g20, eu28, ldc, au) to filter for implementers in the sample. Default: World (as in implemented by one).}

\item{implementer.role}{Bilateral trade flows can be affected by multiple actors. Specify which actor's interventions you want to include. There are three roles: importer, exporter and 3rd country. Combinations are permissible. Default: c('importer','3rd country').}

\item{announcement.period}{Specify a period in which the announcements for your analysis have been made. Default is 'any'. Provide vectors c(after.date, before.date) in R's date format. Also, specify c(after.date, NA) to focus on interventions announced since 'after.date'.}

\item{implementation.period}{Specify a period in which the interventions for your analysis have been implemented. Default is 'any' (incl. not implemented to date). Provide vectors c(after.date, before.date) in R's date format. Also, specify c(after.date, NA) to focus on interventions implemented since 'after.date'.}

\item{revocation.period}{Specify a period in which the interventions for your analysis have been revoked. Default is 'any' (incl. not revoked). Provide vectors c(after.date, before.date) in R's date format. Also, specify c(after.date, NA) to focus on interventions revoked since 'after.date'.}

\item{in.force.today}{Specify whether you want to focus on interventions in force today ('TRUE') or no longer in force today ('FALSE'). Default is 'any'.}

\item{intervention.type}{Specify the names of the trade policy instruments for your analysis. Default is 'any'. For the permissible values, please see the GTA website or the GTA handbook.}

\item{keep.type}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated intervention type.}

\item{mast.chapter}{Specify the MAST chapter IDs for your analysis. Default is 'any'. Permissible values are the MAST chapter letters plus 'tariff', 'fdi', 'migration' and combinations thereof.}

\item{keep.mast}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated MAST chapter ID.}

\item{implementation.level}{Specify the government level responsible for the intervention.  Default is 'any'. Permissible values are 'supranational', 'national', 'subnational', 'IFI', 'NFI' or combinations thereof. IFI and NFI refer to government-owned financial institutions that are owned by one ('NFI') or more ('IFI') governments.}

\item{keep.level}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated implementation levels.}

\item{eligible.firms}{Concentrate the analysis on interventions that are targeted at certain subsets. Default is 'any'. Permissible values are 'all', 'firm-specific', 'SMEs', 'state-controlled','state trading enterprise' or combinations thereof.}

\item{keep.firms}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated firm subsets.}

\item{cpc.sectors}{Provide a vector of CPC codes that you are interested in (version 2.1, any digit level).}

\item{keep.cpc}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated CPC codes.}

\item{hs.codes}{Provide a vector of HS codes that you are interested in (2012 vintage, any digit level).}

\item{keep.hs}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated HS codes.}

\item{intervention.ids}{Provide a vector of intervention IDs.}

\item{keep.interventions}{Specify whether to focus on ('TRUE') or exclude ('FALSE') the stated intervention IDs.}

\item{lag.adjustment}{Create a snapshot of the GTA data at the same point in each calendar year since 2009. Specify a cut-off date ('MM-DD').}

\item{rdata}{Takes value TRUE or FALSE. If TRUE, Rdata file will be stored alongside xlsx. Default: FALSE}

\item{output.path}{Takes the value of the output path (without the filename) added to the working directory as a string starting with "/". Default: None.}

\item{testmode}{Can be set to TRUE or FALSE. If TRUE, sample will be minimized by 99 percent in order to speed up processing. Keep in mind that this will distort results significantly. Default: FALSE.}
}
\value{
Outputs a table with coverage shares ranging from 2009 to 2018 for each importer, exporter, implementer, instrument combination.
}
\description{
Computes the trade coverages of GTA measures.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
