% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta update library.R
\name{gta_update_library}
\alias{gta_update_library}
\title{Get the latest gtalibrary version from GitHub.}
\usage{
gta_update_library(branch = master)
}
\description{
Syncs your local library with our latest GTA GitHub release.
}
