% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs.names}
\alias{hs.names}
\title{HS Names}
\format{An object of class \code{data.frame} with 5205 rows and 2 columns.}
\source{
Global Trade Alert
}
\usage{
hs.names
}
\description{
Dataframe containing HS12 Codes and their names
}
\keyword{datasets}
