% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data description.r
\docType{data}
\name{hs_vintages}
\alias{hs_vintages}
\title{hs_vintage conversion matrix}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6555 rows and 6 columns.
}
\usage{
This data set is used in the function \code{gta_hs_vintage_converter()} to convert
HS codes of any origin into 2012 vintage.
}
\description{
hs_vintages contains the conversion matrix between different HS versions
The underlying conversion matrices are obtained from https://unstats.un.org/unsd/classifications/Econ and allow you
to convert HS codes from one vintage year to another.
}
\keyword{datasets}
