% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.r
\docType{data}
\name{world_geo}
\alias{world_geo}
\title{world_geo}
\format{
The Dataset further contains the following information as retreived from https://unstats.un.org/unsd/methodology/m49/overview/
\describe{
\item{region}{main region (Americas, Europe, Asia etc...)}
\item{sub_region}{Latin America, Western Europe, Eastern Europe etc...}
\item{intermediate_retion}{}
\item{Breakdown between region and sub_region (only defined for some countries)}{Southern Africa, Western Africa, etc...}
}
}
\usage{
world_geo
}
\description{
world_geo is a list which contains 1:The coordinates of the world maps retreived from \code{ggplot2::map_data("world")}.
The country names are adjusted to match the country names used by GTA (see \code{gtalibrary::country.correspondence})
2: The latitute and longtitute for the center of the biggest landmass of each country.
}
\keyword{datasets}
