% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cpc code expand.R
\name{gta_cpc_code_expand}
\alias{gta_cpc_code_expand}
\title{Expands 2-digit level CPC codes into 3-digit level CPC codes.}
\usage{
gta_cpc_code_expand(codes)
}
\arguments{
\item{codes}{Supply the 2-digit level CPC codes you want to expand.}
}
\description{
This function takes in 2-digit level CPC codes and returns all corresponding 3-digit level CPC codes in a vector.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
