% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta data slicer.R
\name{gta_data_slicer}
\alias{gta_data_slicer}
\title{This function allows you to extract a tailored subset of the GTA database.}
\usage{
gta_data_slicer(data.path = "data/master_plus.Rdata",
  gta.evaluation = NULL, affected.flow = NULL,
  implementing.country = NULL, keep.implementer = NULL,
  affected.country = NULL, keep.affected = NULL,
  announcement.period = NULL, implementation.period = NULL,
  revocation.period = NULL, in.force.today = NULL,
  intervention.types = NULL, keep.type = NULL, mast.chapters = NULL,
  keep.mast = NULL, implementation.level = NULL, keep.level = NULL,
  eligible.firms = NULL, keep.firms = NULL, cpc.sectors = NULL,
  keep.cpc = NULL, hs.codes = NULL, keep.hs = NULL,
  intervention.ids = NULL, keep.intervention = NULL,
  lag.adjustment = NULL, df.name = "master.sliced",
  pc.name = "parameter.choice.slicer")
}
\arguments{
\item{data.path}{Specifies where the GTA data file is located (Default: 'data/master_plus.Rdata'). Set to 'online' to download the latest copy.}

\item{gta.evaluation}{Specify what GTA evaluations to include. Default is 'any'. Permissible values are 'red', 'amber', 'green' or combinations thereof.}

\item{affected.flow}{Specify the direction of the trade flow that is affected. The point of view is from the implementing country. Default is 'any'. Permissible values are 'inward', 'outward', 'outward subsidy' or combinations thereof.}

\item{implementing.country}{Specify the implementing countries for your analysis. Default is 'any'. Permissible values are country names or UN codes.}

\item{keep.implementer}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated implementing country.}

\item{affected.country}{Specify the affected countries for your analysis. Default is 'any'. Permissible values are country names or UN codes.}

\item{keep.affected}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated affected country.}

\item{announcement.period}{Specify a period in which the announcements for your analysis have been made. Default is 'any'. Provide vectors c(after.date, before.date) in R's date format. Also, specify c(after.date, NA) to focus on interventions announced since 'after.date'.}

\item{implementation.period}{Specify a period in which the interventions for your analysis have been implemented. Default is 'any' (incl. not implemented to date). Provide vectors c(after.date, before.date) in R's date format. Also, specify c(after.date, NA) to focus on interventions implemented since 'after.date'.}

\item{revocation.period}{Specify a period in which the interventions for your analysis have been revoked. Default is 'any' (incl. not revoked). Provide vectors c(after.date, before.date) in R's date format. Also, specify c(after.date, NA) to focus on interventions revoked since 'after.date'.}

\item{in.force.today}{Specify whether you want to focus on interventions in force today ('TRUE') or no longer in force today ('FALSE'). Default is 'any'.}

\item{intervention.types}{Specify the names of the trade policy instruments for your analysis. Default is 'any'. For the permissible values, please see the GTA website or the GTA handbook.}

\item{keep.type}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated intervention type.}

\item{mast.chapters}{Specify the MAST chapter IDs for your analysis. Default is 'any'. Permissible values are the MAST chapter letters plus 'tariff', 'fdi', 'migration' and combinations thereof.}

\item{keep.mast}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated MAST chapter ID.}

\item{implementation.level}{Specify the government level responsible for the intervention.  Default is 'any'. Permissible values are 'supranational', 'national', 'subnational', 'IFI', 'NFI' or combinations thereof. IFI and NFI refer to government-owned financial institutions that are owned by one ('NFI') or more ('IFI') governments.}

\item{keep.level}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated implementation levels.}

\item{eligible.firms}{Concentrate the analysis on interventions that are targeted at certain subsets. Default is 'any'. Permissible values are 'all', 'firm-specific', 'SMEs', 'state-controlled','state trading enterprise' or combinations thereof.}

\item{keep.firms}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated firm subsets.}

\item{cpc.sectors}{Provide a vector of CPC codes that you are interested in (version 2.1, any digit level).}

\item{keep.cpc}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated CPC codes.}

\item{hs.codes}{Provide a vector of HS codes that you are interested in (2012 vintage, any digit level).}

\item{keep.hs}{Specify whether to focus on ('TRUE') or exclude ('FALSE') interventions with the stated HS codes.}

\item{intervention.ids}{Provide a vector of intervention IDs.}

\item{keep.intervention}{Specify whether to focus on ('TRUE') or exclude ('FALSE') the stated intervention IDs.}

\item{lag.adjustment}{Create a snapshot of the GTA data at the same point in each calendar year since 2009. Specify a cut-off date ('MM-DD').}

\item{df.name}{Set the name of the generated result data frame. Default is master.sliced.}

\item{pc.name}{Set the name of the generated parameter choice data frame. Default is parameter.choice.slicer.}
}
\description{
Make use of the many variables captured in the taxonomy of the Global Trade Alert and tailor the data to your needs. This function returns a list of 2 data frames. The first, master, contains the subset of GTA data you requested. The second, parameter.choices, lists all implicit or exlicited choices that led to the subset.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
