% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta hs code finder.R
\name{gta_hs_code_finder}
\alias{gta_hs_code_finder}
\title{Find HS codes for product names based on the search engines from EU customs and Eurostat.}
\usage{
gta_hs_code_finder(products)
}
\arguments{
\item{products}{A vector of product names (in characters).}
}
\description{
This function looks for HS codes using the search engines of EU customs and Eurostat. For Eurostat, it only collects the result if it is a direct match (i.e. no further information besides the product name has to be provided.). For the EU customs page, it captures all 6-digit HS codes that the search results for the given product name produce.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
