% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta parameter check.R
\name{gta_parameter_check}
\alias{gta_parameter_check}
\title{This is a helper function that evaluates whether all user input is understood. Returns 'OK' if this is so and the unconforming values otherwise.}
\usage{
gta_parameter_check(check.vector = NULL, permissible.values = NULL)
}
\arguments{
\item{check.vector}{Specify the vector you want to check.}

\item{permissible.values}{Specify the permissible values.}
}
\description{
This is a helper function that evaluates whether all user input is understood. Returns 'OK' if this is so and the unconforming values otherwise.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
