% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta parameter check.R
\name{gta_parameter_check}
\alias{gta_parameter_check}
\title{Helper function that evaluates whether user input matches permissible values}
\usage{
gta_parameter_check(
     parameter,
     permissible_values,
     arg_name = NULL,
     warning = FALSE
)
}
\arguments{
\item{parameter}{Specify the argument you want to check.}

\item{permissible_values}{Specify the permissible values.}

\item{arg_name}{Optional specification of the argument's name that should be printed in the error message.}

\item{warning}{specifies whether the ocurrence of a value in parameter which is not contained in permissible_values should give a warning or an error
gives an error by default (ie. Warning = FALSE)}
}
\description{
Helper function that evaluates whether user input matches permissible values
}
\examples{
parameter <- c("one", "TWO", "tHrEE")
permissible_values <- c("one", "two")
# if you do not specify the arg_name, tolower(parameter) will be printed as the arg name.
# To print the passed, unmodified name of the argument, specify: arg_name = parameter
gta_parameter_check(tolower(parameter), permissible_values)
}
