% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta plot map.R
\name{gta_plot_map}
\alias{gta_plot_map}
\title{Create ggplot maps.}
\usage{
gta_plot_map(data = NULL, countries = NULL, value = NULL,
  colour.low = gta_colour$blue[4], colour.high = gta_colour$blue[1],
  range.split = waiver(), legend.title = NULL,
  legend.labels = waiver(), title = NULL)
}
\arguments{
\item{data}{Specify the dataframe with the relevant columns.}

\item{countries}{Specify the name of the column containing the countries.}

\item{value}{Specify the name of the column containing the value for each country.}

\item{colour.low}{Specifiy a color from the gta colour palette for the low part of the gradient. Default: GTA dark blue}

\item{colour.high}{Specify a color from the gta colour palette for the high part of the gradient. Default: GTA light blue}

\item{range.split}{Specify the number of parts the value range should be split to. Either as integer (range will be automatically divided and sequenced) or as sequence (e.g. seq(200,1000,200)). Default: ggplot automatic.}

\item{legend.title}{Specify a title for the colour legend. Default: No title.}

\item{legend.labels}{Specify the legend labels as list. Default: ggplot automatic.}

\item{title}{Specify a title for the map plot. Default: No title.}
}
\description{
This function facilitates map creation with ggplot.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
