% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta setwd.R
\name{gta_setwd}
\alias{gta_setwd}
\title{Sets your working directory to that used by the GTA data team.}
\usage{
gta_setwd(gdrive = NA)
}
\value{
A common working directory
}
\description{
Sets your working directory to that used by the GTA data team.
}
