% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta theme.R
\name{gta_theme}
\alias{gta_theme}
\title{Complete theme for GTA Plots}
\usage{
gta_theme(
   base_size = 12,
   base_family = "Open Sans",
   base_line_size = base_size / 22,
   base_rect_size = base_size / 22,
   aspect_ratio = NULL
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{aspect_ratio}{Here, you can define the aspect ratio of the panel if needed (default is NULL)}
}
\description{
This function defines the GTA theme for ggplot. Add it at the end of a ggplot object.
If you want to overwrite certain elements, simply add \code{theme()} to your plot.
}
\examples{
# Create a plot with gta_theme()
df |>
    ggplot(aes(x = x, y = y, color = color)) +
    geom_line() +
    gta_theme() + # add the base theme to your plot
    theme(axis.text.x = element_text(angle = 45, color = "red", size = 20)) # if you need to deviate from the theme, simply add changes within a new theme() element
}
