% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data description.r
\docType{data}
\name{world_geo}
\alias{world_geo}
\title{coordinates for countries}
\format{
An object of class \code{list} of length 2.
}
\usage{
world_geo
}
\description{
world_geo is a list which contains:
1. the coordinates for all countries to plot a worldmap. Additionally, information on the region, sub_region and intermediate_region
as retreived from https://unstats.un.org/unsd/methodology/m49/overview/ is provided. This allows for easily plotting only certain regions
of the world
2. The latitude and longtitude for the center of the biggest landmass for each country. This data is helpful if eg. labels should be plotted
in the center of a country.
}
\keyword{datasets}
